api/requirements.txt
fastapi[all]==0.112.2
beanie==1.26.0
pymongo==4.9.1
motor==3.6.0
numpy==2.1.0
pytest-asyncio==0.24.0
scipy==1.14.1
minio==7.2.8
redis==5.0.8
celery==5.4.0
python-jose[cryptography]==3.3.0
